[Refactor] FollowFromInsight 인사이트 통계로 옮기기
