+++ keewe-domain/src/main/java/ccc/keewedomain/service/insight/query/InsightStatisticsQueryDomainService.java
@@ -1,5 +1,6 @@
 package ccc.keewedomain.service.insight.query;
 
+import ccc.keewedomain.persistence.repository.insight.FollowFromInsightQueryRepository;
 import ccc.keewedomain.persistence.repository.insight.ProfileVisitFromInsightQueryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -8,8 +9,13 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class InsightStatisticsQueryDomainService {
     private final ProfileVisitFromInsightQueryRepository profileVisitFromInsightQueryRepository;
+    private final FollowFromInsightQueryRepository followFromInsightQueryRepository;
 
     public Long countByInsightId(Long insightId) {
         return profileVisitFromInsightQueryRepository.countByInsightId(insightId);
+    }
+
+    public Long countFollowFromInsightByInsightId(Long insightId) {
+        return followFromInsightQueryRepository.countByInsightId(insightId);
     }
 }
